requirements/local.txt
asgiref
Django
django-model-utils
psycopg2
sqlparse
Unipath